reduce dependency of concerned with rocksdb
