+++ src/components/finutils/src/bins/stt/stt.rs
@@ -24,7 +24,7 @@ use {
             check_delegation_amount, td_addr_to_bytes, BLOCK_INTERVAL, FRA,
             FRA_PRE_ISSUE_AMOUNT,
         },
-        store::utils::fra_gen_initial_tx,
+        utils::fra_gen_initial_tx,
     },
     ruc::*,
     serde::Serialize,
