+++ src/ledger/src/lib.rs
@@ -10,8 +10,12 @@
 pub mod data_model;
 pub mod converter;
 pub mod staking;
-#[cfg(not(target_arch = "wasm32"))]
+pub mod utils;
+
+#[cfg(all(not(target_arch = "wasm32"), feature = "fin_storage"))]
 pub mod store;
+
+pub use fbnc;
 
 use ruc::*;
 const LSSED_VAR: &str = "LEDGER_STATE_SNAPSHOT_ENTRIES_DIR";
