+++ src/ledger/src/store/test.rs
@@ -4,10 +4,12 @@
 use {
     super::{helpers::*, *},
     crate::data_model::{
-        AssetRules, AssetTypeCode, IssueAsset, IssueAssetBody, Memo, Operation,
-        Transaction, TransferAsset, TransferAssetBody, TxOutput, TxnEffect, TxoRef,
-        TxoSID, ASSET_TYPE_FRA, BLACK_HOLE_PUBKEY, TX_FEE_MIN,
+        AssetRules, AssetTypeCode, IssueAsset, IssueAssetBody, IssuerKeyPair, Memo,
+        Operation, Transaction, TransferAsset, TransferAssetBody, TransferType,
+        TxOutput, TxnEffect, TxoRef, TxoSID, ASSET_TYPE_FRA, BLACK_HOLE_PUBKEY,
+        TX_FEE_MIN,
     },
+    crate::utils::{self, *},
     rand_core::SeedableRng,
     zei::{
         anon_xfr::{keys::AXfrKeyPair, structs::OpenAnonAssetRecordBuilder},
